requirements.txt
streamlit>=1.25.0
google-genai>=0.10.0
streamlit-folium>=0.11.0
folium>=0.14.0
fpdf>=1.7.2
pandas
requests
python-dotenv
